🚚 move faceit strategy into a folder
